requirements.txt
lxml==4.9.0
paho_mqtt==1.6.1
pycryptodome==3.15.0
PyYAML==6.0
requests==2.25.1
sslpsk==1.0.0
uptime==3.0.1
urllib3==1.26.5
websocket_client==1.4.1
